% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodatenzentrum.R
\docType{data}
\name{geodatenzentrum}
\alias{geodatenzentrum}
\title{geodatenzentrum}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
geodatenzentrum
}
\description{
tbd.
}
\keyword{datasets}
