% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodatenzentrum.R
\docType{data}
\name{geodatenzentrum}
\alias{geodatenzentrum}
\title{geodatenzentrum}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
tbd.
}
\section{Construction}{
\preformatted{GDZ <- geodatenzentrum$new(local_file_root)
}
\itemize{
\item \code{GDZ}: \code{geodatenzentrum} object.
\item \code{local_file_root}: Path to sync remote data sources to.
}
}

\section{Methods}{

\code{$sync(source)}\cr
Syncs a remote data source to the \code{local_file_root}. Available data sources are listed in \code{$sources}.

\code{$files(collection)}\cr
Lists all synced files for a collection/data source.

\code{$shp_layers(collection)}\cr
Lists all available shp layers for a collection/data source.

\code{$describe(collection)}\cr
Returns information on a collection/data source, e.g. documentation, citation and license.'

\code{$read_sf(collection, layer)}\cr
Returns an object of class \code{sf} for a layer of a collection/data source.

\code{$make_sp(sf)}\cr
Retuns an object of class \code{Spatial} for a given object of class \code{sf}.

\code{$make_df(sp)}\cr
Fortifies a given object of class \code{Spatial} to a \code{data.frame}.

\code{$fortify(collection, layer)}\cr
Combines \code{$read_sf()}, \code{$make_sp()} and \code{$make_df()} and returns a fortified \code{data.frame} for a given layer of a collection/data source.
}

\examples{
\dontrun{
library(geodatenzentrum)

# Create an instance of class geodatenzentrum
GDZ <- geodatenzentrum$new(local_file_root="./geodatenzentrum")

# List available data sources
GDZ$sources

# Sync a data source
GDZ$sync(source = "VG250")

# List available shp layers for the collection
GDZ$shp_layers(collection = "VG250")

# Fortify a shp layer to a data.frame
GDZ$fortify(collection = "VG250", layer = "VG250_KRS")
}
}
\keyword{datasets}
